Move builders to the correct project
AB#9589
